fix(tune): carry the loop's last scoring as f32, matching what the evaluators return
The scoring closures return f32 (`eval_valid` is
`Result<Option<f32>, Box<dyn Error>>`), so annotating the carried pair as
`(f64, Option<f64>)` failed to compile in three places.

Caught by CI rather than locally: this machine is under a build freeze, so the
only local check available was `rustfmt`, which parses but does not type-check.
That is the whole class of error a parse check cannot see, and it is the reason
the previous commit said in as many words that it had not been compiled.

+++ crates/tune/src/train_support/full_driver.rs
@@ -1073,7 +1073,7 @@ pub fn run(config: FullDriverConfig) -> Result<FullDriverOutcome, Box<dyn std::e
     // just scored, recomputing an identical value from an unchanged state: a full train pass plus a
     // full held-out pass, 558s at the sizes above, paid once per run for nothing. Carry the loop's
     // own last scoring forward instead of repeating it.
-    let mut last_scored: Option<(f64, Option<f64>)> = None;
+    let mut last_scored: Option<(f32, Option<f32>)> = None;
     let tstep = Instant::now();
     for step in 1..=steps {
         let tone = Instant::now();
